moving files and code around to create initial login screen, UI is now a mess
