+++ image_processor_app/src/css/ImageProcessor.css
@@ -1,0 +1,36 @@
+.container {
+}
+
+.App-header {
+  background-color: #daf0ee;
+  height: 100px;
+  padding: 20px;
+  color: black;
+  margin: auto;
+}
+
+.App-title {
+  font-size: 2em;
+  font-family: 'Raleway', sans-serif;
+}
+
+.image-container {
+  padding-bottom: 10px;
+}
+
+.basic-text {
+  font-family: 'Raleway', sans-serif;
+}
+
+.footer-text {
+  font-family: 'Raleway', sans-serif;
+  background-color: #daf0ee;
+  margin-top: 50px;
+}
+
+.toggle {
+  padding-top: 10px;
+  padding-bottom: 20px;
+  margin: auto;
+  width: 18%;
+}
