+++ image_processor_app/src/css/LoginScreen.css
@@ -1,0 +1,3 @@
+.container {
+  display: flex;
+}
